requirements.txt
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
PyPDF2
python-dotenv
transformers